- Added method `assign_thumbnail` for `HeifContext` structure. - **BREAKING**: Added default feature `use-bindgen` to control that type of   binding will be used by `libheif-sys` crate (pre-generated or   generated on the fly by `bindgen`) - Fixed creating instance of `ColorProfileType` from instance of   `heif_color_profile_type` on Windows.

+++ src/image_handle.rs
@@ -270,7 +270,8 @@ impl ImageHandle {
             result.set_len(size);
         }
         let c_profile_type = unsafe { lh::heif_image_handle_get_color_profile_type(self.inner) };
-        let profile_type = ColorProfileType::from(c_profile_type);
+        // `c_profile_type` on Windows will be i32, so we need to cast it to u32
+        let profile_type = ColorProfileType::from(c_profile_type as u32);
 
         Some(ColorProfileRaw {
             typ: profile_type,
